requirements.txt
anyio==3.6.1
async-generator==1.10
attrs==21.4.0
bcrypt==3.2.2
beanie==1.11.6
black==22.6.0
bpyutils==0.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
injector==0.20.1
motor==3.0.0
multidict==6.0.2
mypy-extensions==0.4.3
outcome==1.2.0
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
pipupgrade==1.12.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pydantic-yaml==0.8.0
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-dotenv==0.5.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.8
semver==2.13.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
-e git+ssh://git@github.com/cedrata/fastapi-auth-template.git@bc6912b5a51341b213cce674d8e8d71c159abdec#egg=src
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
types-Deprecated==1.2.9
typesentry==0.2.7
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
wrapt==1.14.1
yarl==1.7.2
